Add files via upload
I hope someone finds this code and will prove helpful to him/her. I'll revise this so as to fit the rules in github and or add a bit of documentation. The code however is (I believe) self explanatory so feel free to modify it and perhaps add your modifications here so that other will also benefit.